+++ src/ts/Ads/Managers/CampaignRefreshManager.ts
@@ -22,6 +22,7 @@ import { Diagnostics } from 'Core/Utilities/Diagnostics';
 import { NativePromoEventHandler } from 'Promo/EventHandlers/NativePromoEventHandler';
 import { PromoCampaign } from 'Promo/Models/PromoCampaign';
 import { PurchasingUtilities } from 'Promo/Utilities/PurchasingUtilities';
+import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
 
 export class CampaignRefreshManager extends RefreshManager {
     private _platform: Platform;
@@ -352,6 +353,14 @@ export class CampaignRefreshManager extends RefreshManager {
         if(refreshDelay > 0) {
             this._refillTimestamp = Date.now() + refreshDelay * 1000;
             this._core.Sdk.logDebug('Unity Ads ad plan will expire in ' + refreshDelay + ' seconds');
+
+            if(CustomFeatures.isTimerExpirationExperiment(this._clientInfo.getGameId())) {
+                setTimeout(() => {
+                    if(this._focusManager.isAppForeground()) {
+                        this.refresh();
+                    }
+                }, refreshDelay * 1000 + 1);
+            }
         }
     }